requirements.txt
Flask==3.0.3
beautifulsoup4==4.12.3
requests==2.32.3
google-api-python-client==2.130.0
python-dotenv==1.0.1
Werkzeug==3.0.3
lxml==5.2.2
gunicorn==22.0.0
